% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{admin.censor}
\alias{admin.censor}
\title{Apply administrative censoring to follow up times at a fixed censor.time}
\usage{
admin.censor(follow.up.time, time.var, event.var, censor.time = Inf)
}
\arguments{
\item{follow.up.time}{See \code{\link{SDRsurv}}}

\item{time.var}{See \code{\link{SDRsurv}}}

\item{event.var}{See \code{\link{SDRsurv}}}

\item{censor.time}{Administrative right-censoring time. Default is \code{Inf}, i.e., no censoring}
}
\value{
A data frame with the same shape as \code{follow.up.time} with times and event indicators modified to reflect right-censoring at \code{censor.time}.
}
\description{
Right-censor all observations in \code{follow.up.time} at \code{censor.time}. This is useful when fitting survival curves within each time window
}
