% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clip_interval}
\alias{clip_interval}
\title{Clip all elements in an array to fall in an interval}
\usage{
clip_interval(x, lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{numeric array}

\item{lower}{lower end of the interval}

\item{upper}{upper end of the interval}
}
\value{
clipped array
}
\description{
Clip all elements in an array to fall in an interval
}
\details{
For each element \code{X} in \code{x}, \code{clip_interval(X)} equals \code{lower} if \code{X}<\code{lower}, \code{upper} if \code{X}>\code{upper}, and \code{X} otherwise. This function may be useful to clip predictions to fall in, e.g., \eqn{[0,1]}.
}
