% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDRreg.R
\name{DRtransform}
\alias{DRtransform}
\title{Doubly robust transformation}
\usage{
DRtransform(
  follow.up.time,
  pred_event_censor_obj,
  tvals,
  next.check.in.time = Inf,
  id.var,
  time.var,
  event.var,
  denom.survival.trunc
)
}
\arguments{
\item{follow.up.time}{see \code{\link{SDRsurv}}}

\item{pred_event_censor_obj}{a \code{pred_event_censor} object in the time window of interest}

\item{tvals}{see \code{\link{SDRsurv}}. Must be sorted in ascending order and all greater than the smallest time in \code{pred_event_censor_obj}}

\item{next.check.in.time}{the next check-in time. Default is \code{Inf}, corresponding to the last time window}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{denom.survival.trunc}{see \code{\link{SDRsurv}}}
}
\value{
a named matrix of transformations. Each row corresponds to an individual; each column corresponds to a value of \code{tvals}. Row names are elements in \code{follow.up.time$id.var}; column names are values of \code{tvals}.
}
\description{
Given a \code{pred_event_censor} object in a time window, calculates the doubly robust transformation in the time window. The transformation is used as the outcome when estimating the conditional survival probability at the next check-in time.
}
\section{Warning}{

This function is designed to be called by other functions such as \code{\link{SDRsurv}}, therefore inputs are not thoroughly checked. Incorrect inputs may lead to errors with non-informative messages. The user may call this function if more flexibility is desired.
}

