% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estQ.R
\name{estQ.SuperLearner}
\alias{estQ.SuperLearner}
\title{Estimate (conditional) survival probabilities given pseudo-outcomes with SuperLearner}
\usage{
estQ.SuperLearner(
  covariates,
  stagewise.pseudo.outcomes,
  truncation.index,
  id.var,
  Q.formula = ~.,
  est.cdf = TRUE,
  Q.SuperLearner.control = list(SL.library = "SL.lm")
)
}
\arguments{
\item{covariates}{See \code{\link{SDRsurv}}.}

\item{stagewise.pseudo.outcomes}{a list of list of data frames. Each element in the outer layer corresponds to a value t in \code{tvals} for which to estimate the conditional survival probability. Each element in the inner layer corresponds to a data frame at a check-in time before t. Each data frame should contain a variable to identify each individual (see \code{id.var} argument of \code{\link{SDRsurv}}) and a variable named \code{pseudo.outcome} that contains the pseudo-outcome (e.g., doubly robust transformation) for that check-in time. For check-in times before the truncation time, \code{pseudo.outcome} should equal to 0.}

\item{truncation.index}{index of the check-in time to which left-truncation is applied. The truncation time is \code{check.in.times[truncation.index]}. Covariates available up to (inclusive) \code{check.in.times[truncation.index]} are of interest. See \code{\link{SDRsurv}}}

\item{id.var}{(character) name of the variable that identifies each individual.}

\item{Q.formula}{formula to specify covariates being used for estimating P(T <= t | T > \code{check.in.times[truncation.index]}, covariates available at \code{check.in.times[truncation.index]}). Set to include intercept only (\code{~ 0} or \code{~ -1}) for marginal survival probability, which is simply the mean of pseudo-outcomes. Default is \code{~ .}, which includes main effects of all available covariates up to (inclusive) the \code{truncation.time}.}

\item{est.cdf}{whether to estimate (conditional) CDF or survival function. Default is \code{TRUE}. Set to \code{FALSE} to estimate the survival function.}

\item{Q.SuperLearner.control}{a list containing optional arguments passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}}. We encourage using a named list. Will be passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} by running a command like \code{do.call(SuperLearner, Q.SuperLearner.control)}. The user should not specify \code{Y}, \code{X} and \code{family}. This argument is ignored if \code{Q.formula} has no covariates. When \code{Q.formula} only includes an intercept, \code{\link[SuperLearner]{SuperLearner}} is not called.}
}
\value{
a list of fitted \code{SuperLearner} models corresponding to each t in \code{tvals}. If \code{Q.formula} is empty, then return a list of numbers, each being estimated P(T <= t) for t in \code{tvals}.
}
\description{
Estimate P(T <= t | T > truncation time, covariates available at truncation time) for given t, where T is the time to event.
}
\section{Warning}{

This function is designed to be called by other functions such as \code{\link{SDRsurv}}, therefore inputs are not thoroughly checked. Incorrect inputs may lead to errors with non-informative messages. The user may call this function if more flexibility is desired.
}

\section{Custom learners}{

Custom learners may be specified by providing an element named \code{SL.library} in \code{Q.SuperLearner.control}.The user may refer to resources such as \url{https://cran.r-project.org/web/packages/SuperLearner/vignettes/Guide-to-SuperLearner.html} for a guide to create custom learners.
}

