% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_akritas}
\alias{fit_akritas}
\title{Wrapper of \code{survivalmodels::akritas}}
\usage{
fit_akritas(
  formula,
  data,
  id.var,
  time.var,
  event.var,
  nfold = 1,
  option = list(),
  lambda = 0.5,
  ...
)
}
\arguments{
\item{formula}{formula used by \code{\link[survivalmodels:akritas]{survivalmodels::akritas}}}

\item{data}{data containing all covariates, follow-up time, event indicator and id}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{option}{a list containing optional arguments passed to \code{\link[survivalmodels:deepsurv]{survivalmodels::akritas}}. We encourage using a named list. Will be passed to \code{\link[survivalmodels:akritas]{survivalmodels::akritas}} by running a command like \code{do.call(akritas, option)}. The user should not specify \code{formula}, \code{data} and \code{reverse}; \code{time_variable}, \code{status_variable}, \code{x}, \code{y} will be ignored.}

\item{lambda}{bandwidth parameter for uniform smoothing kernel in nearest neighbours estimation. The default value of 0.5 is arbitrary and should be chosen by the user}

\item{...}{ignored}
}
\value{
a \code{\link{pred_surv}} class containing fitted survival curves for individuals in \code{data}
}
\description{
Wrapper of \code{survivalmodels::akritas}
}
