% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_ctree}
\alias{fit_ctree}
\title{Wrapper of \code{party::ctree}}
\usage{
fit_ctree(
  formula,
  data,
  id.var,
  time.var,
  event.var,
  nfold = 1,
  option = list(),
  ...
)
}
\arguments{
\item{formula}{formula used by \code{\link[party]{ctree}}}

\item{data}{data containing all covariates, follow-up time, event indicator and id}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{option}{a list containing optional arguments passed to \code{\link[party]{ctree}}. We encourage using a named list. Will be passed to \code{\link[party]{ctree}} by running a command like \code{do.call(ctree, option)}. The user should not specify \code{formula} and \code{data}.}

\item{...}{ignored}
}
\value{
a \code{\link{pred_surv}} class containing fitted survival curves for individuals in \code{data}
}
\description{
Wrapper of \code{party::ctree}
}
