% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_dnnsurv}
\alias{fit_dnnsurv}
\title{Wrapper of \code{survivalmodels::dnnsurv}}
\usage{
fit_dnnsurv(
  formula,
  data,
  id.var,
  time.var,
  event.var,
  nfold = 1,
  option = list(),
  ...
)
}
\arguments{
\item{formula}{formula used by \code{\link[survivalmodels:dnnsurv]{survivalmodels::dnnsurv}}}

\item{data}{data containing all covariates, follow-up time, event indicator and id}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{option}{a list containing optional arguments passed to \code{\link[survivalmodels:deepsurv]{survivalmodels::dnnsurv}}. We encourage using a named list. Will be passed to \code{\link[survivalmodels:dnnsurv]{survivalmodels::dnnsurv}} by running a command like \code{do.call(dnnsurv, option)}. The user should not specify \code{formula}, \code{data} and \code{reverse}; \code{time_variable}, \code{status_variable}, \code{x}, \code{y} will be ignored.}

\item{...}{ignored}
}
\value{
a \code{\link{pred_surv}} class containing fitted survival curves for individuals in \code{data}
}
\description{
Wrapper of \code{survivalmodels::dnnsurv}
}
