% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_rpart}
\alias{fit_rpart}
\title{Wrapper of \code{rpart::rpart} and \code{partykit::as.party}}
\usage{
fit_rpart(
  formula,
  data,
  id.var,
  time.var,
  event.var,
  nfold = 1,
  option = list(),
  ...
)
}
\arguments{
\item{formula}{formula used by \code{\link[rpart:rpart]{rpart::rpart}}}

\item{data}{data containing all covariates, follow-up time, event indicator and id}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{option}{a list containing optional arguments passed to \code{\link[rpart:rpart]{rpart::rpart}}. We encourage using a named list. Will be passed to \code{\link[rpart:rpart]{rpart::rpart}} by running a command like \code{do.call(rpart, option)}. The user should not specify \code{formula} and \code{data}.}

\item{...}{ignored}
}
\value{
a \code{\link{pred_surv}} class containing fitted survival curves for individuals in \code{data}
}
\description{
First use \code{\link[rpart:rpart]{rpart::rpart}} to obtain a tree and then use \code{\link[partykit:party-coercion]{as.party}} to obtain Kaplan-Meier fits.
}
