% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_surv_option}
\alias{fit_surv_option}
\title{Options of machine learning methods' wrappers for fitting conditional survival curves}
\usage{
fit_surv_option(nfold = 1, option = list(), oob = TRUE)
}
\arguments{
\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{option}{a list containing optional arguments passed to the wrapped machine learning function. Will be used in a command like \code{do.call(machine.learning, option)}.}

\item{oob}{whether to use out-of-bag (OOB) fitted values from forests when sample splitting is not used (\code{nfold=1}). Ignored for trees or \code{coxph}.}
}
\description{
Options of machine learning methods' wrappers for fitting conditional survival curves
}
