% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_survival_forest}
\alias{fit_survival_forest}
\title{Wrapper of \code{grf::survival_forest}}
\usage{
fit_survival_forest(
  formula,
  data,
  id.var,
  time.var,
  event.var,
  nfold = 1,
  oob = TRUE,
  option = list(),
  ...
)
}
\arguments{
\item{formula}{formula used by \code{\link[grf:survival_forest]{grf::survival_forest}}}

\item{data}{data containing all covariates, follow-up time, event indicator and id}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{oob}{whether to use out-of-bag (OOB) fitted values from \code{\link[grf:survival_forest]{grf::survival_forest}} when sample splitting is not used (\code{nfold=1}). Default is \code{TRUE}}

\item{option}{a list containing optional arguments passed to \code{\link[grf:survival_forest]{grf::survival_forest}}. We encourage using a named list. Will be passed to \code{\link[grf:survival_forest]{grf::survival_forest}} by running a command like \code{do.call(survival_forest, option)}. The user should not specify \code{X}, \code{Y}, \code{D}, \code{failure.times} and \code{compute.oob.predictions}.}

\item{...}{ignored}
}
\value{
a \code{\link{pred_surv}} class containing fitted survival curves for individuals in \code{data}
}
\description{
Wrapper of \code{grf::survival_forest}
}
