% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surv.R
\name{fit_survSuperLearner}
\alias{fit_survSuperLearner}
\title{Wrapper of \code{survSuperLearner::survSuperLearner}}
\usage{
fit_survSuperLearner(
  formula,
  data,
  id.var,
  time.var,
  event.var,
  nfold = 1,
  option = list(event.SL.library = c("survSL.coxph", "survSL.weibreg", "survSL.gam",
    "survSL.rfsrc"), cens.SL.library = c("survSL.coxph", "survSL.weibreg", "survSL.gam",
    "survSL.rfsrc")),
  ...
)
}
\arguments{
\item{formula}{formula containing all covariates to be used}

\item{data}{data containing all covariates, follow-up time, event indicator and id}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{nfold}{number of folds used when fitting survival curves with sample splitting. Default is 1, meaning no sample splitting}

\item{option}{a list containing optional arguments passed to \code{\link[survSuperLearner:survSuperLearner]{survSuperLearner::survSuperLearner}}. We encourage using a named list. Will be passed to \code{\link[survSuperLearner:survSuperLearner]{survSuperLearner::survSuperLearner}} by running a command like \code{do.call(survSuperLearner, option)}. The user should not specify \code{time}, \code{event}, \code{X}, \code{newX} or \code{new.times}. We encourage the user to specify \code{event.SL.library} and \code{cens.SL.library}.}

\item{...}{ignored}
}
\value{
a \code{\link{pred_event_censor}} class containing fitted survival curves for individuals in \code{data}
}
\description{
Wrapper of \code{survSuperLearner::survSuperLearner}
}
