% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greg.R
\name{Greg.SuperLearner}
\alias{Greg.SuperLearner}
\title{Regression based on G-computation transformation of fitted survival and censoring probabilities using SuperLearner}
\usage{
Greg.SuperLearner(
  covariates,
  follow.up.time,
  pred_event.list,
  check.in.times,
  tvals,
  truncation.index,
  id.var,
  time.var,
  event.var,
  Q.formula = ~.,
  Q.SuperLearner.control = list(family = gaussian(), SL.library = "SL.lm")
)
}
\arguments{
\item{covariates}{see \code{\link{SDRsurv}}}

\item{follow.up.time}{see \code{\link{SDRsurv}}}

\item{pred_event.list}{list of \code{pred_surv} objects for time to event. Each \code{pred_event_censor} object in the list corresponds to a time window in \code{check.in.times} after \code{truncation.index} in increasing order.}

\item{check.in.times}{see \code{\link{SDRsurv}}}

\item{tvals}{see \code{\link{SDRsurv}}. Must be sorted in ascending order.}

\item{truncation.index}{see \code{\link{SDRsurv}}}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}

\item{Q.formula}{formula to specify covariates being used for estimating P(T > t | T > \code{check.in.times[truncation.index]}, covariates available at \code{check.in.times[truncation.index]}). Set to include intercept only (\code{~ 0} or \code{~ -1}) for marginal survival probability, which is simply the mean of pseudo-outcomes. Default is \code{~ .}, which includes main effects of all available covariates up to (inclusive) the \code{truncation.time}.}

\item{Q.SuperLearner.control}{see \code{\link{SDRsurv}}}
}
\value{
a list of \code{\link{mult_stage_survfit}} objects, each corresponding to a value in \code{tvals}
}
\description{
Apply G-computation transformation on fitted survival and censoring probabilities in each time window and estimate P(T > t | T > truncation time, covariates available at truncation time) with \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}}.
}
\section{Warning}{

This function is designed to be called by other functions such as \code{\link{SDRsurv}}, therefore inputs are not thoroughly checked. Incorrect inputs may lead to errors with non-informative messages. The user may call this function if more flexibility is desired.
}

\section{Custom learners}{

Custom learners may be specified by providing an element named \code{SL.library} in \code{Q.SuperLearner.control}.The user may refer to resources such as \url{https://cran.r-project.org/web/packages/SuperLearner/vignettes/Guide-to-SuperLearner.html} for a guide to create custom learners.
}

