% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{intercept_model}
\alias{intercept_model}
\title{S3 class for predictive models with an intercept only}
\usage{
intercept_model(value)
}
\arguments{
\item{value}{the predicted mean value}
}
\value{
an "\code{intercept_model}" object, essentially a list with element \code{value}.
}
\description{
S3 class for predictive models with an intercept only
}
