% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWsurv.R
\name{IPWsurv}
\alias{IPWsurv}
\title{Estimate (conditional) survival probabilities with inverse probability weighting (IPW) transformation}
\usage{
IPWsurv(
  covariates,
  follow.up.time,
  check.in.times,
  tvals = NULL,
  truncation.index = 1,
  id.var,
  time.var,
  event.var,
  censor.formula = NULL,
  Q.formula = ~.,
  censor.method = c("rfsrc", "ctree", "rpart", "cforest", "coxph", "coxtime",
    "deepsurv", "dnnsurv", "akritas"),
  censor.control = fit_surv_option(),
  Q.SuperLearner.control = list(family = gaussian(), SL.library = "SL.lm"),
  denom.survival.trunc = 0.001
)
}
\arguments{
\item{covariates}{a list of data frames of covarates in the order of check in times. Each data frame contains the covariates collected at a check-in time. Data frames may have different numbers of variables (may collect different variables at different check-in times) and different numbers of individuals (some individuals may have an event or is censored before a later check-in time). All data frames must have a common character variable (see \code{id.var}) that identifies each individual but no other variables with common names. No missing data is allowed.}

\item{follow.up.time}{data frame of follow up times, i.e., times to event/censoring. Contains the variable that identifies each individual, the follow up times and an indicator of event/(right-)censoring. Follow up times must be numeric. Indicator of event/censoring should be binary with 0=censored, 1=event.}

\item{check.in.times}{numeric/integer vector of check-in times in ascending order. The first check-in time is typically the baseline.}

\item{tvals}{times t for which P(T > t) given covariates are computed (T is the time to event). Default is all unique event times in \code{follow.up.time}. Will be sorted in ascending order.}

\item{truncation.index}{index of the check-in time to which left-truncation is applied. The truncation time is \code{check.in.times[truncation.index]}. Covariates available up to (inclusive) \code{check.in.times[truncation.index]} are of interest. Default is 1, corresponding to no truncation.}

\item{id.var}{(character) name of the variable that identifies each individual.}

\item{time.var}{(character) name of the variable containing follow up times in the data frame \code{follow.up.time}.}

\item{event.var}{(character) name of the variable containing indicator of event/censoring in the data frame \code{follow.up.time}.}

\item{censor.formula}{a list of formulas to specify covariates being used when estimating the conditional survival probabilities of time to censoring at each check-in time. The length should be the number of check in times after \code{truncation.index} (inclusive). Default is \code{~ .} for all check-in times, which includes main effects of all covariates available at each check-in time.}

\item{Q.formula}{formula to specify covariates being used for estimating P(T > t | T > \code{check.in.times[truncation.index]}, covariates available at \code{check.in.times[truncation.index]}). Set to include intercept only (\code{~ 0} or \code{~ -1}) for marginal survival probability. Default is \code{~ .}, which includes main effects of all available covariates up to (inclusive) the \code{check.in.times[truncation.index]}.}

\item{censor.method}{one of \code{"rfsrc"}, \code{"ctree"}, \code{"rpart"}, \code{"cforest"}, \code{"coxph"}, \code{"coxtime"}, \code{"deepsurv"}, \code{"dnnsurv"}, \code{"akritas"}. The machine learning method to fit survival survival curves of time to censoring in each time window. Default is \verb{"rfsrc}. See the underlying wrappers \code{\link{fit_rfsrc}}, \code{\link{fit_ctree}}, \code{\link{fit_rpart}}, \code{\link{fit_cforest}}, \code{\link{fit_coxph}}, \code{\link{fit_coxtime}}, \code{\link{fit_deepsurv}}, \code{\link{fit_dnnsurv}}, \code{\link{fit_akritas}} for the available options.}

\item{censor.control}{a returned value from \code{\link{fit_surv_option}}}

\item{Q.SuperLearner.control}{a list containing optional arguments passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}}. We encourage using a named list. Will be passed to \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} by running a command like \code{do.call(SuperLearner, Q.SuperLearner.control)}. Default is \code{list(SL.library="SL.lm")}, which uses linear regression. The user should not specify \code{Y} and \code{X}, and must specify \code{SL.library} if default is not used. If \code{family} is gaussian by default if unspecified, and must be gaussian if specified, with a possibly non-identity link. When \code{Q.formula} only includes an intercept, \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} will not be called and the default setting can be used.}

\item{denom.survival.trunc}{the numeric truncation value for the survival function in the denominator. All denominators below \code{denom.survival.trunc} will be set to \code{denom.survival.trunc} for numerical stability.}
}
\value{
a list of fitted \code{SuperLearner} models corresponding to each t in \code{tvals}.
}
\description{
Estimate P(T > t | T > truncation time, covariates available at truncation time) for given t, where T is the time to event, using inverse probability weighting (IPW) transformation. Use a user-specified flexible method to fit survival curves of time to event/censoring at each stage and then use \code{\link[SuperLearner:SuperLearner]{SuperLearner::SuperLearner}} to regress pseudo-outcome on covariates in order to estimate P(T > t | T > truncation time, covariates available at truncation time).
}
\section{Formula arguments}{

All formulas should have covariates on the right-hand side and no terms on the left-hand side, e.g., \code{~ V1 + V2 + V3}. At each check-in time, the corresponding formulas may (and usually should) contain covariates at previous check-in times, and must only include available covariates up to (inclusive) that check-in time. Interactions, polynomials and splines may be treated differently by different machine learning methods to estimate conditional survival curves.
}

