% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv_class.R
\name{pred_event_censor}
\alias{pred_event_censor}
\title{Constructor of S3 class \code{pred_event_censor}}
\usage{
pred_event_censor(pred_event_obj, pred_censor_obj = NULL)
}
\arguments{
\item{pred_event_obj}{a \code{pred_surv} object holding predicted/fitted survival probabilities for events.}

\item{pred_censor_obj}{a \code{pred_surv} object holding predicted/fitted survival probabilities for censoring. Set to \code{NULL} (default) if no censoring present.}
}
\description{
Construct a \code{pred_event_censor} object containing predicted/fitted survival and censoring probabilities for multiple individuals in the same sample.  Can only hold survival curves of the form of decreasing right-continuous step functions.
}
\details{
The predicted/fitted probability matrices in \code{pred_event_obj} and \code{pred_censor_obj} must have same row names, i.e., same number of individuals.
}
