% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedsurv.R
\name{pred_surv}
\alias{pred_surv}
\title{Constructor of S3 class \code{pred_surv}}
\usage{
pred_surv(time, surv)
}
\arguments{
\item{time}{collection of all unique event times where the predicted/fitted survival function may have a jump. Must be increasing.}

\item{surv}{matrix of predicted/fitted survival probabilities. Each row corresponds to an individual; each column corresponds to an event time in \code{time}. Row names should be \code{id.var} that identifies each individual (see \code{\link{SDRsurv}}).}
}
\description{
Construct a \code{pred_surv} object containing predicted/fitted survival probabilities for multiple individuals in the same sample. Can only hold survival curves of the form of decreasing rightcontinuous step functions.
}
\details{
If there is no event, set \code{time=Inf} and \code{surv} to be a matrix with only one column of 1.
}
