% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{predict.mult_stage_survfit}
\alias{predict.mult_stage_survfit}
\alias{fitted.mult_stage_survfit}
\title{Predict and fitted method for mult_stage_survfit}
\usage{
\method{predict}{mult_stage_survfit}(
  object,
  newdata = object$covariate.data,
  pred.prob.trunc = c(0, 1),
  ...
)

\method{fitted}{mult_stage_survfit}(object, pred.prob.trunc = c(0, 1), ...)
}
\arguments{
\item{object}{a "\code{mult_stage_survfit}" object}

\item{newdata}{new data frame of covariates used for prediction. Default is the \code{covariate.data} stored in \code{object}}

\item{pred.prob.trunc}{values at which to truncate predicted survival probabilities from each model in \code{object}. Must be a numeric vector with 2 numbers in \eqn{[0,1]}. Default is \code{c(0,1)}.}

\item{...}{further arguments passed to \code{predict} methods of the models in \code{object}}
}
\value{
a vector of predicted (conditional) survival probabilities. If no covariates are used, that is, interested in the survival probability marginalized over all covariates so that all individuals have the same prediction, then only outputs that one predicted survial probability.
}
\description{
Predict and fitted method for mult_stage_survfit
}
