% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estQ.R
\name{prep.Q.reg.data}
\alias{prep.Q.reg.data}
\title{Prepare data used in regression for estimating (conditional) suvival probabilities}
\usage{
prep.Q.reg.data(
  covariates,
  stagewise.pseudo.outcomes,
  truncation.index,
  id.var
)
}
\arguments{
\item{covariates}{See \code{\link{SDRsurv}}.}

\item{stagewise.pseudo.outcomes}{a list of list of data frames. Each element in the outer layer corresponds to a value t in \code{tvals} for which to estimate the conditional survival probability. Each element in the inner layer corresponds to a data frame at a check-in time (i.e., stage) before t. Each data frame should contain a variable to identify each individual (see \code{id.var} argument of \code{\link{SDRsurv}}) and a variable named \code{pseudo.outcome} that contains the pseudo-outcome (e.g., doubly robust transformation) for that check-in time. For check-in times before the truncation time, \code{pseudo.outcome} should equal to 0.}

\item{truncation.index}{index of the check-in time to which left-truncation is applied. The truncation time is \code{check.in.times[truncation.index]}. Covariates available up to (inclusive) \code{check.in.times[truncation.index]} are of interest. See \code{\link{SDRsurv}}}

\item{id.var}{name of the variable that identifies each individual. Just provide the expression such as \code{id} and do not embrace it with \code{""} or \code{''}.}
}
\value{
A list containing:
\enumerate{
\item \code{pseudo.outcomes}: a list of data frames with pseudo-outcomes. Each data frame corresponds to one t in \code{tvals} (see \code{\link{SDRsurv}}) and contains two variables: a variable with same name as \code{id.var}; a variable named \code{pseudo.outcome}
\item \code{history}: a data frame containing all historic variables
}
All data frames are sorted by \code{id.var}.
}
\description{
Sum up pseudo-outcomes in each time window and produce the final pseudo-outcome in the regression to estimate P(T <= t | T > truncation time, covariates available at truncation time); link covariates at each check-in time up to truncation index to form the data with all historic covariates that can be used in the regression. The output can be used by \code{\link[SuperLearner]{SuperLearner}} or other regression methods for the final estimation.
}
\section{Warning}{

This function is designed to be called by other functions such as \code{\link{SDRsurv}}, therefore inputs are not thoroughly checked. Incorrect inputs may lead to errors with non-informative messages. The user may call this function if more flexibility is desired.
}

