% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv_class.R
\name{truncate_pred_surv}
\alias{truncate_pred_surv}
\title{Truncate \code{pred_surv} object to minimal that covers a time window}
\usage{
truncate_pred_surv(pred_surv_obj, end = Inf)
}
\arguments{
\item{pred_surv_obj}{\code{pred_surv} object.}

\item{end}{ending point (inclusive) of at which truncation applies. Default is \code{Inf}.}
}
\value{
A \code{pred_surv} object with truncated times.
}
\description{
Truncate the time and survival probabilities in a \code{pred_surv} object to minimal so that predicted/fitted survival probability at \code{end} is kept. This function can be used before applying transformations (e.g., \code{\link{DRtransform}}, \code{\link{Gtransform}}, \code{\link{IPWtransform}}) to avoid unnecessary computation. No truncation is applied if \code{end} is greater than or equal to the last time in the \code{pred_surv} object.
}
