% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWtransform.R
\name{IPWtransform}
\alias{IPWtransform}
\title{Inverse probability weighting (IPW) transformation of fitted survival and censoring probabilities}
\usage{
IPWtransform(
  follow.up.time,
  pred_censor.list,
  check.in.times,
  tvals,
  truncation.index,
  id.var,
  time.var,
  event.var
)
}
\arguments{
\item{follow.up.time}{see \code{\link{SDRsurv}}}

\item{pred_censor.list}{list of \code{pred_surv} objects (see \code{\link{pred_surv}}) containing survival probabilities for time to censoring. Each \code{pred_surv} object in the list corresponds to a time window in \code{check.in.times} after \code{truncation.index} in increasing order.}

\item{check.in.times}{see \code{\link{SDRsurv}}}

\item{tvals}{see \code{\link{SDRsurv}}. Must be sorted in ascending order.}

\item{truncation.index}{see \code{\link{SDRsurv}}}

\item{id.var}{see \code{\link{SDRsurv}}}

\item{time.var}{see \code{\link{SDRsurv}}}

\item{event.var}{see \code{\link{SDRsurv}}}
}
\value{
a list of list of data frames containing doubly robust transformed pseudo-outcomes that can be used by \code{\link{estQ.SuperLearner}}.
}
\description{
Apply inverse probability weighting (IPW) transformation on fitted survival and censoring probabilities in each time window for estimating P(T <= t | T > truncation time, covariates available at truncation time).
}
\section{Warning}{

This function is designed to be called by \code{\link{IPWsurv}}, therefore inputs are not thoroughly checked. Incorrect inputs may lead to errors with non-informative messages. The user may call this function if more flexibility is desired.
}

