% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{mult_stage_survfit}
\alias{mult_stage_survfit}
\title{S3 class for the predictive model for (conditional) survival probability with multiple stages}
\usage{
mult_stage_survfit(
  covariate.data,
  formula,
  visit.times,
  tval,
  truncation.index,
  models
)
}
\arguments{
\item{covariate.data}{data frame containing the covariates used to train the models}

\item{formula}{formula used in the models}

\item{visit.times}{numeric/integer vector of visit times in ascending order. The first visit time is typically the baseline.}

\item{truncation.index}{index of the visit time to which left-truncation is applied. The truncation time is \code{visit.times[truncation.index]}. Covariates available up to (inclusive) \code{visit.times[truncation.index]} are of interest.}

\item{models}{list of models. Each model can predict the survival probability beyond the next visit time given survival at the current visit time. Models should be in an ascending order according to the corresponding visit times.}

\item{tvals}{time t for which P(T > t) (potentially given covariates) is estimated (T is the time to event)}
}
\value{
a "\code{mult_stage_survfit}" object, essentially a list with input parameters of this function.
}
\description{
This function is the creator of \code{mult_stage_survfit} class. A \code{mult_stage_survfit} object is a collection of fitted models for conditional survival probability within each time window. The associated \code{predict} and \code{fitted} methods autocmatically conduct the prediction, which is multiplying all model predictions, potentially after clipping the predictions into an interval in \eqn{[0,1]}.
}
